pyproject.toml
[tool.poetry]
name = "data-inventory"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-bigquery = "^3.27.0"
pandas = "^2.2.3"
db-dtypes = "^1.3.1"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
